+++ scripts/devops_tasks/setup_execute_tests.py
@@ -22,6 +22,8 @@ dev_setup_script_location = os.path.join(root_dir, 'scripts/dev_setup.py')
 # evaluating whether we want this or not.
 ALLOWED_RETURN_CODES = []
 
+MANAGEMENT_PACKAGE_IDENTIFIERS = ['mgmt', 'azure-cognitiveservices-personalizer']
+
 def prep_and_run_tests(targeted_packages, python_version, test_res):
     print('running test setup for {}'.format(targeted_packages))
     run_check_call([python_version, dev_setup_script_location, '-p', ','.join([os.path.basename(p) for p in targeted_packages])], root_dir)
@@ -29,14 +31,14 @@ def prep_and_run_tests(targeted_packages, python_version, test_res):
     # if we are targeting only packages that are management plane, it is a possibility 
     # that no tests running is an acceptable situation
     # we explicitly handle this here.
-    if all(map(lambda x : 'mgmt' in x, targeted_packages)):
+    if all(map(lambda x : any([pkg_id in x for pkg_id in MANAGEMENT_PACKAGE_IDENTIFIERS]), targeted_packages)):
         ALLOWED_RETURN_CODES.append(5)
 
     print('Setup complete. Running pytest for {}'.format(targeted_packages))
     command_array = [python_version, '-m', 'pytest']
     command_array.extend(test_res)
     command_array.extend(targeted_packages)
-    run_check_call(command_array, root_dir, ALLOWED_RETURN_CODES)
+    run_check_call(command_array, root_dir, ALLOWED_RETURN_CODES, True)
 
 if __name__ == '__main__':
     parser = argparse.ArgumentParser(description = 'Install Dependencies, Install Packages, Test Azure Packages, Called from DevOps YAML Pipeline')
@@ -58,6 +60,12 @@ if __name__ == '__main__':
         dest='test_results',
         help=('The folder where the test results will be stored in xml format.'
               'Example: --junitxml="junit/test-results.xml"'))
+
+    parser.add_argument(
+        '--mark_arg',
+        dest='mark_arg',
+        help=('The complete argument for `pytest -m "<input>"`. This can be used to exclude or include specific pytest markers.'
+              '--mark_arg="not cosmosEmulator"'))
 
     parser.add_argument(
         '--disablecov',
@@ -86,5 +94,8 @@ if __name__ == '__main__':
 
     if args.disablecov:
         test_results_arg.append('--no-cov')
+
+    if args.mark_arg:
+        test_results_arg.extend(['-m', '"{}"'.format(args.mark_arg)])
 
     prep_and_run_tests(targeted_packages, args.python_version, test_results_arg)