+++ sdk/cosmos/azure-cosmos/test/retry_policy_tests.py
@@ -30,6 +30,7 @@ from azure.cosmos.http_constants import HttpHeaders, StatusCodes, SubStatusCodes
 import azure.cosmos.retry_utility as retry_utility
 import test_config
 
+pytestmark = pytest.mark.cosmosEmulator
 
 #IMPORTANT NOTES: