+++ sdk/hanaonazure/azure-mgmt-hanaonazure/azure/mgmt/hanaonazure/models/os_profile.py
@@ -24,18 +24,23 @@ class OSProfile(Model):
     :vartype os_type: str
     :ivar version: Specifies version of operating system.
     :vartype version: str
+    :ivar ssh_public_key: Specifies the SSH public key used to access the
+     operating system.
+    :vartype ssh_public_key: str
     """
 
     _validation = {
         'computer_name': {'readonly': True},
         'os_type': {'readonly': True},
         'version': {'readonly': True},
+        'ssh_public_key': {'readonly': True},
     }
 
     _attribute_map = {
         'computer_name': {'key': 'computerName', 'type': 'str'},
         'os_type': {'key': 'osType', 'type': 'str'},
         'version': {'key': 'version', 'type': 'str'},
+        'ssh_public_key': {'key': 'sshPublicKey', 'type': 'str'},
     }
 
     def __init__(self, **kwargs):
@@ -43,3 +48,4 @@ class OSProfile(Model):
         self.computer_name = None
         self.os_type = None
         self.version = None
+        self.ssh_public_key = None