+++ sdk/keyvault/azure-security-keyvault/tests/test_examples_keyvault.py
@@ -26,7 +26,7 @@ def create_vault_client():
     )
 
     # Create a new Vault client using Azure credentials
-    vault_client = VaultClient(vault_url=vault_url, credentials=credentials)
+    vault_client = VaultClient(vault_url=vault_url, credential=credentials)
     # [END create_vault_client]
     return vault_client
 
@@ -46,7 +46,7 @@ def create_secret_client():
     )
 
     # Create a new Secret client using Azure credentials
-    secret_client = SecretClient(vault_url=vault_url, credentials=credentials)
+    secret_client = SecretClient(vault_url=vault_url, credential=credentials)
     # [END create_secret_client]
     return secret_client
 
@@ -95,14 +95,14 @@ class TestExamplesKeyVault(KeyVaultTestCase):
             pass
 
         try:
-            # [START update_secret_attributes]
+            # [START update_secret]
 
             # update attributes of an existing secret
 
             content_type = "text/plain"
             tags = {"foo": "updated tag"}
             secret_version = secret.version
-            updated_secret = secret_client.update_secret_attributes(
+            updated_secret = secret_client.update_secret(
                 "secret-name", secret_version, content_type=content_type, tags=tags
             )
 
@@ -111,7 +111,7 @@ class TestExamplesKeyVault(KeyVaultTestCase):
             print(updated_secret.content_type)
             print(updated_secret.tags)
 
-            # [END update_secret_attributes]
+            # [END update_secret]
         except HttpResponseError:
             pass